[#33] set up page routes and linking
