+++ content/home.yml
@@ -56,6 +56,6 @@ tableOfContents:
     description: >-
       Pictures I've painted and sketched and composited and imagined.
   - heading: Writing
-    link: writing
+    link: writing/page/1
     description: >-
       Thoughts on, well, anything that strikes me.
