--- src/routes/writing/index.json.js
@@ -1,14 +1,0 @@
-import generateWritingList from '@/utils/generateWritingList.js'
-
-export function get(req, res) {
-	res.writeHead(200, {
-		'Content-Type': 'application/json'
-	})
-
-	res.end(JSON.stringify(
-    generateWritingList('content/writing', {
-      start: 0,
-      end: 10
-    })
-  ))
-}
